+++ lib/core/utils/frame-messenger.js
@@ -1,5 +1,5 @@
-import { post } from './respondable/post';
-import { messageHandler } from './respondable/message-handler';
+import { post } from './frame-messenger/post';
+import { messageHandler } from './frame-messenger/message-handler';
 
 let handler;
