+++ lib/core/utils/frame-messenger/assert-window.js
