+++ lib/core/utils/frame-messenger/message-parser.js
